% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath_utils.R
\name{absolute_path}
\alias{absolute_path}
\title{Convert relative paths to absolute paths}
\usage{
absolute_path(path, wd = getwd())
}
\arguments{
\item{path}{the path to check and convert to absolute}

\item{wd}{path to prepend to a relative path. Current working directory by default.}
}
\value{
absolute path
}
\description{
First checks if path is relative or absolute, and if relative prepends wd (or other given absolute path)
}
\keyword{internal}
