% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodelist_edgelist_gen.R
\name{add_interaction}
\alias{add_interaction}
\title{Add an interaction line to the interaction edgelist}
\usage{
add_interaction(
  edgelist,
  agent,
  object,
  agent_ident = c("friend", "person"),
  agent_ident_prob = c(0.7, 0.3),
  object_ident = c("friend", "person"),
  object_ident_prob = c(0.7, 0.3),
  institution = NA,
  rseed = NA
)
}
\arguments{
\item{edgelist}{interaction edgelist to add to}

\item{agent}{string agent's name}

\item{object}{string object's name}

\item{agent_ident}{string or string list: agent's understanding of their own identity}

\item{agent_ident_prob}{numeric or numeric list summing to 1: probability of taking each identity in agent_ident}

\item{object_ident}{string or string list: agent's understanding of object's identity}

\item{object_ident_prob}{numeric or numeric list summing to 1: probability of taking each identity in object_ident}

\item{institution}{string or string list: the institution(s) identities in this relationship can come from}

\item{rseed}{the random seed}
}
\value{
edgelist data frame with added interaction line
}
\description{
If an interaction is given optional arguments (institution, rseed) that are not already columns in the provided dataframe, those columns are added.
}
