% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_file.R
\name{agent}
\alias{agent}
\title{Write agent into input}
\usage{
agent(
  agenttext,
  name,
  dictname = "usfullsurveyor2015",
  dicttype = "mean",
  dictgender = "av",
  impression_eqns = "us2010",
  emotion_eqns = "us2010",
  impression_gender = c("neutral", "female", "male"),
  emotion_gender = c("neutral", "female", "male"),
  eqndir = "/Users/aidan/Desktop/interact_data/eq_coefs/clean"
)
}
\arguments{
\item{agenttext}{file template to add agent information to}

\item{name}{string agent name}

\item{dictname}{string, key of an available dictionary OR filepath}

\item{dicttype}{string (\code{"mean"}, \code{"sd"}, or \code{"cov"})}

\item{dictgender}{string (\code{"av"}, \code{"male"}, or \code{"female"})}

\item{impression_eqns}{string, key of an available equation set OR filepath}

\item{emotion_eqns}{string, key of an available equation set OR filepath}

\item{impression_gender}{string (\code{"neutral"}, \code{"female"}, or \code{"male"})}

\item{emotion_gender}{string (\code{"neutral"}, \code{"female"}, or \code{"male"})}

\item{eqndir}{string, path to directory where equations are located}
}
\value{

}
\description{
Compile all necessary information for an individual agent and add it to a template object that can be written to a text file
}
\keyword{internal}
