% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_checks.R
\name{check_abbrev}
\alias{check_abbrev}
\title{Check abbreviation validity}
\usage{
check_abbrev(value, allowed)
}
\arguments{
\item{value}{the entry to check}

\item{allowed}{the list of allowed abbreviations/alternate spellings}
}
\value{
logical indicating success
}
\description{
Check abbreviation validity
}
