% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_checks.R
\name{check_agent_opt_args}
\alias{check_agent_opt_args}
\title{Check agent optional argument input}
\usage{
check_agent_opt_args(opt_args)
}
\arguments{
\item{opt_args}{named vector of provided optional arguments}
}
\value{
logical for successful check
}
\description{
institution, alphas, betas, deltas, numsamples
}
\keyword{internal}
