% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_checks.R
\name{check_dict_components}
\alias{check_dict_components}
\title{Compatibility check: dictionary and component}
\usage{
check_dict_components(dictname, indices = c(1, 2, 3, 4))
}
\arguments{
\item{dictname}{length 4 list, order: identity, behavior, identity, mod}
}
\value{
boolean for successful check
}
\description{
this checks that the provided dictionary does indeed exist for the required component (identity, behavior, modifier).
}
\keyword{internal}
