% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_utils.R
\name{check_dict_gender}
\alias{check_dict_gender}
\title{Compatibility check: dictionary and gender}
\usage{
check_dict_gender(dictname, gender, dicts)
}
\arguments{
\item{dictname}{string}

\item{gender}{string (\code{"av"}, \code{"male"}, \code{"female"})}

\item{dicts}{list of dictionary objects}
}
\value{

}
\description{
this checks that the provided dictionary has the requested gender (if it is an included dict). This will look different when things are provided as a list versus as singular.
}
\keyword{internal}
