% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_checks.R
\name{check_dict_group}
\alias{check_dict_group}
\title{Compatibility check: dictionary and group}
\usage{
check_dict_group(dictname, group, indices = c(1, 2, 3, 4))
}
\arguments{
\item{dictname}{string or length 4 vector}

\item{group}{string or length 4 vector (\code{"all"}, \code{"m"}, \code{"f"})}
}
\value{
boolean for successful check
}
\description{
this checks that the provided dictionary has the requested group (if it is an included dict). This will look different when things are provided as a list versus as singular.
}
\keyword{internal}
