% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_checks.R
\name{check_dict_stat}
\alias{check_dict_stat}
\title{Compatibility check: dictionary and stat}
\usage{
check_dict_stat(dict, dictstat, indices = c(1, 2, 3, 4))
}
\arguments{
\item{dictstat}{string (\code{"mean"}, \code{"sd"}, or \code{"cov"})}

\item{dictname}{string}
}
\value{
boolean for successful check
}
\description{
this checks that the provided dictionary has the provided stat. This will look different for dictionaries accessed with keywords than dictionaries provided with a filepath.
}
\keyword{internal}
