% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_checks.R
\name{check_dict_type}
\alias{check_dict_type}
\title{Compatibility check: dictionary and type}
\usage{
check_dict_type(dict, dicttype)
}
\arguments{
\item{dicttype}{string (\code{"mean"}, \code{"sd"}, or \code{"cov"})}

\item{dictname}{string}
}
\value{
boolean for successful check
}
\description{
this checks that the provided dictionary has the provided type. This will look different for dictionaries accessed with keywords than dictionaries provided with a filepath.
}
\keyword{internal}
