% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_utils.R
\name{check_dict_type_compatibility}
\alias{check_dict_type_compatibility}
\title{Compatibility check: dictionary and type}
\usage{
check_dict_type_compatibility(dictname, dicttype, dicts)
}
\arguments{
\item{dictname}{string}

\item{dicttype}{string (\code{"mean"}, \code{"sd"}, or \code{"cov"})}

\item{dicts}{list of dictionary objects}
}
\value{

}
\description{
this checks that the provided dictionary has the provided type. This will look different for dictionaries accessed with keywords than dictionaries provided with a filepath.
}
\keyword{internal}
