% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_checks.R
\name{check_eqn_gender}
\alias{check_eqn_gender}
\title{Check that specified equations and equation genders are compatible}
\usage{
check_eqn_gender(eqns, eqns_gender)
}
\arguments{
\item{eqns}{equation list}

\item{eqns_gender}{equation gender list}
}
\value{
boolean successful check
}
\description{
Check that specified equations and equation genders are compatible
}
