% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_checks.R
\name{check_events}
\alias{check_events}
\title{Check events file input for format errors}
\usage{
check_events(events)
}
\arguments{
\item{events}{dataframe}
}
\value{
boolean successful check
}
\description{
Check events file input for format errors
}
