% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bayesact.R
\name{check_exit}
\alias{check_exit}
\title{Check whether BayesACT completed successfully and print a warning if not.}
\usage{
check_exit(termId, outfile)
}
\arguments{
\item{termId}{the terminal process id}

\item{outfile}{path to terminal output file}
}
\description{
Check whether BayesACT completed successfully and print a warning if not.
}
