% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_checks.R
\name{check_identity_prob_match}
\alias{check_identity_prob_match}
\title{Check that all identities have a corresponding probability and vice versa}
\usage{
check_identity_prob_match(ident, prob)
}
\arguments{
\item{ident}{identity list}

\item{prob}{probability list}
}
\value{
logical for successful check
}
\description{
Are lists the same length?
}
\keyword{internal}
