% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_checks.R
\name{check_input_list}
\alias{check_input_list}
\title{Check input for agreement with provided lists}
\usage{
check_input_list(input, allowlist, allowlength, allowsingle, allowfile)
}
\arguments{
\item{input}{vector}

\item{allowlist}{vector}

\item{allowlength}{numeric}

\item{allowsingle}{boolean}

\item{allowfile}{boolean}
}
\value{
boolean true if successful check
}
\description{
It checks that filepaths are valid, but does not check whether the file contents are as required. I assume BayesACT does this?
}
\keyword{internal}
