% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_checks.R
\name{check_interaction_opt_args}
\alias{check_interaction_opt_args}
\title{Check institution optional arguments}
\usage{
check_interaction_opt_args(opt_args)
}
\arguments{
\item{opt_args}{named list of optional arguments}
}
\value{
logical for successful check
}
\description{
Check institution optional arguments
}
\keyword{internal}
