% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_checks.R
\name{check_probs}
\alias{check_probs}
\title{Check that probabilities given sum to 1}
\usage{
check_probs(agent_ident_prob, object_ident_prob)
}
\arguments{
\item{agent_ident_prob}{list of agent probabilities}

\item{object_ident_prob}{list of object probabilities}
}
\value{
boolean successful test
}
\description{
Check that probabilities given sum to 1
}
\keyword{internal}
