% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodelist_edgelist_gen.R
\name{component_split}
\alias{component_split}
\title{Split a given data frame into identity, behavior, and modifier components, if possible. Return error if not.}
\usage{
component_split(df)
}
\arguments{
\item{df}{given object}
}
\value{
list of length 4--identity, behavior, identity, modifier
}
\description{
Split a given data frame into identity, behavior, and modifier components, if possible. Return error if not.
}
\keyword{internal}
