% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_datafile.R
\name{construct_df_filename}
\alias{construct_df_filename}
\title{Given a key and assorted information OR a dataframe, construct and return a sensible file name.}
\usage{
construct_df_filename(
  df = NA,
  key = NA,
  gender = NA,
  component = NA,
  stat = NA
)
}
\arguments{
\item{df}{dataframe object}

\item{key}{key string}

\item{gender}{gender string}

\item{component}{component string}

\item{stat}{stat string}
}
\value{
string with the filename
}
\description{
This will be in format "key_gender_component_stat.csv" if the dict was provided via actdata key,
or "dfname.csv" if the dict was provided as a dataframe.
}
\details{
These are all dictionaries and so should be csvs
}
