% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath_utils.R
\name{create_dir_if_needed}
\alias{create_dir_if_needed}
\title{Create directory if it doesn't already exist}
\usage{
create_dir_if_needed(dir)
}
\arguments{
\item{dir}{path to directory}
}
\value{
logical whether directory needed to be created or not
}
\description{
Create directory if it doesn't already exist
}
\keyword{internal}
