% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_utils.R
\name{dict_specs}
\alias{dict_specs}
\title{Checks whether an entry is a valid file, a valid actdata key, or a correctly formatted dataset. Throws an error otherwise.}
\usage{
dict_specs(dict)
}
\arguments{
\item{dict}{object to test}
}
\value{
character. file, key, df.
}
\description{
Checks whether an entry is a valid file, a valid actdata key, or a correctly formatted dataset. Throws an error otherwise.
}
