% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_file.R
\name{event_lines}
\alias{event_lines}
\title{Create events line for input file}
\usage{
event_lines(events, filepath)
}
\arguments{
\item{events}{a list of events}

\item{filepath}{a filepath to save to}
}
\value{
the line as a character vector
}
\description{
Create events line for input file
}
\keyword{internal}
