% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_file.R
\name{eventfile_out}
\alias{eventfile_out}
\title{Write out event file}
\usage{
eventfile_out(template, filename, dir)
}
\arguments{
\item{template}{event dataframe}

\item{filename}{filename to save under}

\item{dir}{directory name to save in}
}
\value{
filepath saved under
}
\description{
Write out event file
}
