% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_utils.R
\name{expand}
\alias{expand}
\title{Expand string into vector of given length for line spec}
\usage{
expand(s, len)
}
\arguments{
\item{s}{the string/vector}

\item{len}{desired length}
}
\value{
vector of desired length
}
\description{
If object passed is already a vector, no change
}
\keyword{internal}
