% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_format_utils.R
\name{extract_dict_list}
\alias{extract_dict_list}
\title{extract_dict_list}
\usage{
extract_dict_list(input)
}
\arguments{
\item{input}{the input passed from the nodelist df}
}
\value{
a list with four string entries or a tibble with four nested tibble entries
}
\description{
Handles dict input from nodelist tibbles.
This can be:
\enumerate{
\item a list containing one tibble with four rows containing four strings (if actdata keys or filepaths)
\item or a list of four tibbles representing dictionaries (if a df was passed)
}
}
\details{
These are not checked for validity here; that happens later. TODO: Think about whether this check should be moved up (possibly to the nodelist creation).
}
