% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_utils.R
\name{fileinput}
\alias{fileinput}
\title{Path check}
\usage{
fileinput(dictname)
}
\arguments{
\item{dictname}{string to check}
}
\value{
boolean
}
\description{
This checks if a string is a valid file path. It returns a boolean value or vector of boolean values (if vector passed in)
}
\details{
TODO: Where does this look? Will relative paths work?
}
\keyword{internal}
