% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_format_utils.R
\name{get_actor_prob_line}
\alias{get_actor_prob_line}
\title{Utility for building a line in the interaction section of the input file}
\usage{
get_actor_prob_line(identities, probs)
}
\arguments{
\item{identities}{list of identities}

\item{probs}{list of probabilities}
}
\value{
text string to enter in line
}
\description{
Utility for building a line in the interaction section of the input file
}
