% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_format_utils.R
\name{get_agent_opt_arg_lines}
\alias{get_agent_opt_arg_lines}
\title{Get lines to insert in file for agent optional arguments}
\usage{
get_agent_opt_arg_lines(opt_args)
}
\arguments{
\item{opt_args}{named list of optional arguments for one agent}
}
\value{
list of lines for file
}
\description{
Get lines to insert in file for agent optional arguments
}
\keyword{internal}
