% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_utils.R
\name{get_eqn_file}
\alias{get_eqn_file}
\title{Construct file string (equation)}
\usage{
get_eqn_file(eqndir, eqns, gender, type)
}
\arguments{
\item{eqndir}{directory where equations are located}

\item{eqns}{string}

\item{gender}{string (\code{"neutral"}, \code{"male"}, \code{"female"})}

\item{type}{string (\code{"mean"}, \code{"sd"}, \code{"cov"})}
}
\value{

}
\description{
this builds the correct filepath for the equations and checks if the specified gender is available. If the equations are given as a filepath, it returns the filepath and ignores specified gender.
}
\keyword{internal}
