% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_format_utils.R
\name{get_interaction_opt_arg_lines}
\alias{get_interaction_opt_arg_lines}
\title{Get lines to insert in file for interaction optional arguments}
\usage{
get_interaction_opt_arg_lines(opt_args)
}
\arguments{
\item{opt_args}{optional arguments}
}
\value{
list of lines for file
}
\description{
Get lines to insert in file for interaction optional arguments
}
\keyword{internal}
