% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_format_utils.R
\name{get_lists}
\alias{get_lists}
\title{Convert comma-separated strings in cell entries to lists when needed.}
\usage{
get_lists(row)
}
\arguments{
\item{row}{the dataframe row to convert and return}
}
\value{
row with needed entries converted to lists
}
\description{
Useful for when actor or interaction input is read in from a csv file.
}
