% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_format_utils.R
\name{insert_lines}
\alias{insert_lines}
\title{Insert lines into template}
\usage{
insert_lines(file, lines, start, end, insertAt = "end")
}
\arguments{
\item{file}{template object}

\item{lines}{lines to insert}

\item{start}{line after which to start}

\item{end}{line before which to end}

\item{insertAt}{string, whether to put as close as possible to the start or the end (\code{"start"}, \code{"end"})}
}
\value{
file with additional lines
}
\description{
Insert lines into template
}
\keyword{internal}
