% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_file.R
\name{interaction}
\alias{interaction}
\title{Add an interaction line to the simulation text file}
\usage{
interaction(
  agent,
  object,
  agent_ident = "person",
  agent_ident_prob = "1",
  object_ident = "person",
  object_ident_prob = "1",
  opt_args = ""
)
}
\arguments{
\item{agent}{agent name}

\item{object}{object name}

\item{agent_ident}{agent identities}

\item{agent_ident_prob}{agent identity probabilities}

\item{object_ident}{object identities}

\item{object_ident_prob}{object identity probabilities}

\item{opt_args}{more arguments}
}
\value{
the lines to add
}
\description{
Add an interaction line to the simulation text file
}
\keyword{internal}
