% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_utils.R
\name{make_file_string}
\alias{make_file_string}
\title{Construct file string (dictionary)}
\usage{
make_file_string(dir, dict, gender, type)
}
\arguments{
\item{dir}{string directory}

\item{dict}{string}

\item{gender}{string (\code{"av"}, \code{"male"}, \code{"female"})}

\item{type}{string (\code{"mean"}, \code{"sd"}, \code{"cov"})}
}
\value{

}
\description{
this constructs the correct file string from dictionary information (dictionaries only, currently)
}
\keyword{internal}
