% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_file.R
\name{out}
\alias{out}
\title{Write to .txt file}
\usage{
out(
  template,
  filename,
  dir = "/Users/aidan/Desktop/School/Grad_school/inteRact/bayesact/Rscripts/outtest/"
)
}
\arguments{
\item{template}{object to write}

\item{filename}{name to write to}

\item{dir}{directory to write to}
}
\value{

}
\description{
Write to .txt file
}
\keyword{internal}
