% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_format_utils.R
\name{remove_line}
\alias{remove_line}
\title{Remove lines that contain a given string}
\usage{
remove_line(string, file)
}
\arguments{
\item{string}{the marker string}

\item{file}{the file to remove from}
}
\value{
file with lines containing string removed
}
\description{
Remove lines that contain a given string
}
