% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_datafile.R
\name{save_actdata_input}
\alias{save_actdata_input}
\title{Save actdata objects where bayesact can get them}
\usage{
save_actdata_input(dataname)
}
\arguments{
\item{dataname}{name of the dataset in actdata}
}
\value{
filename that the object got saved to
}
\description{
Need to move these files to a folder in the user's working directory so BayesACT can access them.
Make a directory "actdata_dicts_eqns" under the user's working directory (if necessary)
}
