% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_datafile.R
\name{save_dict_df}
\alias{save_dict_df}
\title{Save dataframe dictionary}
\usage{
save_dict_df(data, filename, bayesact_dir)
}
\arguments{
\item{dir}{}
}
\value{

}
\description{
The newest shiniest save function
}
