% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_datafile.R
\name{save_for_bayesact}
\alias{save_for_bayesact}
\title{Save files where bayesact can find them}
\usage{
save_for_bayesact(dataname, class, filepath)
}
\arguments{
\item{dataname}{name of actdata object}

\item{class}{string "dict" or "eqn"}

\item{filepath}{string filepath to save under}
}
\description{
Save files where bayesact can find them
}
