% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_file.R
\name{simfile_out}
\alias{simfile_out}
\title{Write to .txt file}
\usage{
simfile_out(template, filename, dir)
}
\arguments{
\item{template}{object to write}

\item{filename}{name to write to}

\item{dir}{directory to write to}
}
\description{
Write to .txt file
}
\keyword{internal}
