% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_utils.R
\name{standardize_option}
\alias{standardize_option}
\title{standardize_option}
\usage{
standardize_option(input, param, version = "dict")
}
\arguments{
\item{input}{the string to standardize}

\item{param}{the dictionary parameter expected (gender, component, stat)}

\item{version}{dict or eqn}
}
\value{
the standardized version of the input string
}
\description{
This function deals with abbreviations in parameter specification and returns the spellings that are used in the datasets.
}
\keyword{internal}
