% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bayesact.R
\name{wait_until_done}
\alias{wait_until_done}
\title{Pause until terminal process is done}
\usage{
wait_until_done(termId)
}
\arguments{
\item{termId}{the terminal id that is running}
}
\value{
boolean success
}
\description{
Pause until terminal process is done
}
