% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_file.R
\name{write_input_file_from_df}
\alias{write_input_file_from_df}
\title{Write BayesACT input file}
\usage{
write_input_file_from_df(nodelist, edgelist)
}
\arguments{
\item{nodelist}{a dataframe giving dictionary information for each actor}

\item{edgelist}{a dataframe delineating starting parameters (actor identity vector, object identity vector, probabilities) for each dyad}
}
\description{
Write out a text file in the format required by BayesACT containing all necessary information for all actors
}
