% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodelist_edgelist_gen.R
\name{add_actor}
\alias{add_actor}
\title{Add an actor line to a actor nodelist data frame}
\usage{
add_actor(
  nodelist,
  name,
  dict = "usfullsurveyor2015",
  dict_stat = "mean",
  dict_gender = "av",
  eqns = "us2010",
  eqns_gender = c("av", "female"),
  alphas = NA,
  betas = NA,
  deltas = NA,
  numsamples = NA
)
}
\arguments{
\item{nodelist}{data frame to add to}

\item{name}{string actor's name}

\item{dict}{string or string list length 4: dictionary to use. Entries must either be dictionary keys from the actdata package or valid filepaths to dictionaries.}

\item{dict_stat}{string or string list length 4: stat of provided dictionaries (mean, cov, sd). For datasets in actdata, check available stats with actdata::dict_info().}

\item{dict_gender}{string or string list length 4: gender of provided dictionaries (av, female, male). For datasets in actdata, check available genders with actdata::dict_info().}

\item{eqns}{string or string list length 2: equations to use. Entries must either by equation keys from the actdata package or valid filepaths to coefficient matrices.}

\item{eqns_gender}{string or string list length 2: gender of equations to use (av, female, male). For datasets in actdata, check available genders with actdata::eqn_info().}

\item{alphas}{numeric or numeric list length 3: alpha value(s) to use.}

\item{betas}{numeric or numeric list length 2: beta value(s) to use.}

\item{deltas}{numeric or numeric list length 2: delta value(s) to use.}

\item{numsamples}{numeric: number of samples to draw}
}
\value{
provided nodelist with actor line appended
}
\description{
If the actor is given optional arguments (alphas, betas, deltas, numsamples) that are not already columns in the data frame, these columns will be added.
}
