% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_format_utils.R
\name{add_leading_zeros}
\alias{add_leading_zeros}
\title{Add leading zero before decimal point to probs if necessary}
\usage{
add_leading_zeros(probs)
}
\arguments{
\item{probs}{vector of probabilities}
}
\value{
formatted character vector with leading zero before decimal
}
\description{
Add leading zero before decimal point to probs if necessary
}
