% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_file.R
\name{agent}
\alias{agent}
\title{Write agent into input}
\usage{
agent(
  name,
  dict = "usfullsurveyor2015",
  dict_type = "mean",
  dict_gender = "av",
  eqns = "us2010",
  eqns_gender = c("av", "female"),
  opt_args = ""
)
}
\arguments{
\item{name}{string agent name}

\item{dict}{string or length 4 vector, key(s) of an available dictionary OR filepath(s)}

\item{dict_type}{string or length 4 vector (\code{"mean"}, \code{"sd"}, or \code{"cov"})}

\item{dict_gender}{string or length 4 vector (\code{"av"}, \code{"male"}, or \code{"female"})}

\item{eqns}{string or length 2 vector, key(s) of an available equation set OR filepath(s)}

\item{eqns_gender}{string or length 2 vector (\code{"av"}, \code{"f"}, or \code{"m"})}
}
\value{
file with inserted lines
}
\description{
Compile all necessary information for an individual agent and add it to a template object that can be written to a text file
}
\keyword{internal}
