% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodelist_edgelist_gen.R
\name{blank_edgelist}
\alias{blank_edgelist}
\title{Create a blank interaction edgelist}
\usage{
blank_edgelist(use.institution = FALSE, use.rseed = FALSE)
}
\arguments{
\item{use.institution}{boolean whether to include an institution column}

\item{use.rseed}{boolean whether to include an rseed column}
}
\value{
empty edgelist data frame with correct column names
}
\description{
Create a blank interaction edgelist
}
