% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodelist_edgelist_gen.R
\name{blank_nodelist}
\alias{blank_nodelist}
\title{Create a blank actor nodelist in dataframe format}
\usage{
blank_nodelist(
  use.alphas = FALSE,
  use.betas = FALSE,
  use.deltas = FALSE,
  use.numsamples = FALSE
)
}
\arguments{
\item{use.alphas}{boolean whether to include alphas column}

\item{use.betas}{boolean whether to include betas column}

\item{use.deltas}{boolean whether to include deltas column}

\item{use.numsamples}{boolean whether to include numsamples column}
}
\value{
empty nodelist data frame with correct column names
}
\description{
This can be added to manually or with the help of add_actor()
}
