% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_utils.R
\name{check_dict_components}
\alias{check_dict_components}
\title{Compatibility check: dictionary and component}
\usage{
check_dict_components(dictname, dicts)
}
\arguments{
\item{dictname}{string}

\item{dicts}{list of dictionary objects}
}
\value{

}
\description{
this checks that the provided dictionary does indeed exist for the required component (identity, behavior, modifier). This will look different when things are provided as a list versus as singular.
}
\keyword{internal}
