% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_checks.R
\name{check_dict_gender}
\alias{check_dict_gender}
\title{Compatibility check: dictionary and gender}
\usage{
check_dict_gender(dictname, gender)
}
\arguments{
\item{dictname}{string or length 4 vector}

\item{gender}{string or length 4 vector (\code{"av"}, \code{"m"}, \code{"f"})}
}
\value{
boolean for successful check
}
\description{
this checks that the provided dictionary has the requested gender (if it is an included dict). This will look different when things are provided as a list versus as singular.
}
\keyword{internal}
