% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_utils.R
\name{get_eqn_file}
\alias{get_eqn_file}
\title{Construct file string (equation)}
\usage{
get_eqn_file(key, group, component, bayesact_dir)
}
\arguments{
\item{group}{string}

\item{component}{string (\code{"impression"}, \code{"emotion"})}

\item{eqn}{string}
}
\value{
string filepath
}
\description{
this builds the correct filepath for the equations and checks if the specified group is available.
if the equations are given as a filepath, it returns the filepath and ignores specified group.
if the input is a name from actdata, it saves the dataset in the "actdata_dicts_eqns" folder in the working directory
}
\keyword{internal}
