% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_utils.R
\name{make_file_string}
\alias{make_file_string}
\title{Construct file string (dictionary)}
\usage{
make_file_string(dict, gender, component, type)
}
\arguments{
\item{dict}{string}

\item{gender}{string (\code{"av"}, \code{"female"}, \code{"male"})}

\item{component}{string (\code{"identities"}, \code{"behaviors"}, \code{"settings"}, \code{"mods"})}
}
\value{
string filepath
}
\description{
this constructs the correct file string from dictionary information
if the input is a name from actdata, it saves the dataset in the "actdata_dicts_eqns" folder in the working directory
}
\keyword{internal}
