% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_utils.R
\name{make_file_string}
\alias{make_file_string}
\title{Construct file string (dictionary)}
\usage{
make_file_string(dict, spec, key, group, component, stat, bayesact_dir)
}
\arguments{
\item{dict}{string}

\item{group}{string (\code{"all"}, \code{"female"}, \code{"male"})}

\item{component}{string (\code{"identity"}, \code{"behavior"}, \code{"setting"}, \code{"modifier"})}

\item{stat}{string (\code{"mean"}, \code{"sd"}, \code{"cov"})}

\item{bayesact_dir}{string}
}
\value{
string filepath
}
\description{
this constructs the correct file string from dictionary information
if the input is a name from actdata, it saves the dataset in the "actdata_dicts_eqns" folder in the working directory
}
\keyword{internal}
