% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_file.R
\name{out}
\alias{out}
\title{Write to .txt file}
\usage{
out(template, filename, dir = "bayesact_input_files")
}
\arguments{
\item{template}{object to write}

\item{filename}{name to write to}

\item{dir}{directory to write to}
}
\description{
Write to .txt file
}
\keyword{internal}
