% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bayesact.R
\name{run_bayesact}
\alias{run_bayesact}
\title{Run bayesact for a single simulation file}
\usage{
run_bayesact(
  simfilename,
  bayesact_dir,
  input_dir = "bayesact_input",
  output_dir = "bayesact_output",
  wd = getwd()
)
}
\arguments{
\item{simfilename}{the name of the simulation file}

\item{bayesact_dir}{the top level directory where the bayesact code lives}

\item{input_dir}{the directory the file is in}

\item{output_dir}{the directory in which to put the bayesact output}

\item{wd}{current working directory}
}
\value{
output file path
}
\description{
Code adapted from that written by Jon Morgan and Kim Rogers
}
\details{
TODO: There are more options here (see p. 7-8 of bayesactv2.pdf).
}
