% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_datafile.R
\name{save_dict_df}
\alias{save_dict_df}
\title{Save dataframe dictionary}
\usage{
save_dict_df(data, filename, bayesact_dir)
}
\arguments{
\item{data}{data object to save}

\item{filename}{name to save under (and return)}

\item{bayesact_dir}{filepath to bayesact toplevel directory}
}
\value{
file name
}
\description{
The newest shiniest save function
}
\keyword{internal}
