% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_datafile.R
\name{save_eqn_actdata}
\alias{save_eqn_actdata}
\title{#' Save actdata objects where bayesact can get them
#'
#' Save them to a "data" folder that is under the bayesact directory
#'
#' @param dataname name of the dataset in actdata
#' @param bayesact_dir top directory for bayesact code
#'
#' @return filename that the object got saved to
save_actdata_input <- function(dataname, bayesact_dir){}
\usage{
save_eqn_actdata(dataname, bayesact_dir)
}
\arguments{
\item{dataname}{name of actdata object}

\item{filepath}{string filepath to save under}
}
\description{
#' Save actdata objects where bayesact can get them
#'
#' Save them to a "data" folder that is under the bayesact directory
#'
#' @param dataname name of the dataset in actdata
#' @param bayesact_dir top directory for bayesact code
#'
#' @return filename that the object got saved to
save_actdata_input <- function(dataname, bayesact_dir){
}
\details{
create_dir_if_needed(dirpath)

if(grepl("dict", dataname)){
# the object is a dictionary
class <- "dict"
filename <- paste0(dataname, ".csv")
} else {
# the object is an equation set
class <- "eqn"
filename <- paste0(dataname, ".dat")
}

filepath <- file.path(dirpath, filename)

save_for_bayesact(dataname, class = class, filepath = filepath)
return(filename)
}
#' Save files where bayesact can find them
#'
#' @param dataname name of actdata object
#' @param class string "dict" or "eqn"
#' @param filepath string filepath to save under
#'
#' @import actdata
save_for_bayesact <- function(dataname, class, filepath){
}
\section{path <- file.path(getwd(), "actdata_dicts_eqns")}{

}

\section{path <- "/Users/aidan/Desktop/School/Grad_school/bayesactgithub/bayesact/data"}{
dirpath <- file.path(bayesact_dir, "data")
}

\section{TODO I think this can be deleted if save_eqn_bayesact works}{

}

\section{TODO this is SUPER sensitive to the input format of the dataframe. Does it work with every actdata dataset?}{
data <- get(dataname, asNamespace("actdata"))
}

\section{if the dictionary is stat "mean", it needs to have six EPA columns and an institution codes column}{

}

\section{neither seems to be the case for COV and SD datasets but this needs to be checked.}{

}

