% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_datafile.R
\name{save_eqn_actdata}
\alias{save_eqn_actdata}
\title{Save files where bayesact can find them}
\usage{
save_eqn_actdata(data, dataname, bayesact_dir)
}
\arguments{
\item{data}{equation data object (from actdata)}

\item{dataname}{name of actdata object}

\item{bayesact_dir}{top level of bayesact code directory}
}
\description{
Save files where bayesact can find them
}
\keyword{internal}
